Update replxx after author abandoned it

+++ src/libs/replxx/windows.cxx
@@ -99,6 +99,7 @@ T* HandleEsc(HANDLE out_, T* p, T* end) {
 
 int win_write( HANDLE out_, bool autoEscape_, char const* str_, int size_ ) {
 	int count( 0 );
+	/// FIXME: Replace this variable with function call passing an output descriptor as a parameter.
 	if ( tty::out ) {
 		DWORD nWritten( 0 );
 		if ( autoEscape_ ) {
@@ -113,7 +114,7 @@ int win_write( HANDLE out_, bool autoEscape_, char const* str_, int size_ ) {
 						int toWrite( static_cast<int>( str_ - s ) );
 						WriteConsoleA( out_, s, static_cast<DWORD>( toWrite ), &nWritten, nullptr );
 						count += nWritten;
-						if ( nWritten != toWrite ) {
+						if ( static_cast<int>( nWritten ) != toWrite ) {
 							s = str_ = nullptr;
 							break;
 						}
